refactor(ui): split monolithic ui.py into dialog modules
ui.py war auf 1104 Zeilen angewachsen — eine Klasse, die Header, Grid,
drei modale Dialoge, Token-Refresh und Hover-Logik vereint. Die Dialoge
werden zu freistehenden Funktionen extrahiert; App bleibt reiner
Koordinator.

- src/theme.py — Farbpalette, Fonts, TIME_VALUES, PAUSE_VALUES und
  apply_combobox_style zentralisiert (waren bisher Module-Konstanten in
  ui.py). DAYS_DE/MONTHS_DE bleiben in ui.py bzw. report.py, weil
  display-spezifisch.
- src/dialogs/entry_dialog.py — open_entry_dialog(parent, date_str,
  storage, settings, on_change). on_change ist der Refresh-Hook der
  Kalenderansicht.
- src/dialogs/settings_dialog.py — open_settings_dialog mit lokalen
  make_entry/make_text/make_combo Closures gegen die Style-Boilerplate.
- src/dialogs/send_dialog.py — open_send_dialog plus
  show_missing_credentials_dialog. _default_from_date ist eine pure
  Funktion, build_date_row faltet die Von-/Bis-Zeile.
- resolve_autostart_target zog von ui.py nach autostart.py um — sauberer
  Ort für plattformabhängige Autostart-Logik. Test-Import entsprechend
  umgebogen.

Wirkung: ui.py 1104 → 443 Zeilen. Pyflakes sauber, 89/89 Tests grün.
