+++ src/theme.py
@@ -1,0 +1,53 @@
+from tkinter import ttk
+
+# Dark Modern color palette
+BG = "#1a1a2e"
+CELL_BG = "#16213e"
+WEEKEND_BG = "#0f3460"
+ACCENT = "#e94560"
+ACCENT_HOVER = "#c73550"
+STATUS_OK = "#4ade80"
+TEXT = "#e0e0e0"
+TEXT_MUTED = "#888888"
+ENTRY_BG = "#1a3a5c"
+WEEKEND_ENTRY_BG = "#1a3050"
+WEEKEND_FG = "#6c6c80"
+
+FONT = ("Segoe UI", 10)
+FONT_SMALL = ("Segoe UI", 8)
+FONT_BOLD = ("Segoe UI", 10, "bold")
+FONT_HEADER = ("Segoe UI", 16, "bold")
+FONT_FOOTER = ("Segoe UI", 12, "bold")
+
+# Hover colors (slightly lighter variants)
+CELL_BG_HOVER = "#1e2d52"
+WEEKEND_BG_HOVER = "#153a6e"
+ENTRY_BG_HOVER = "#224a70"
+WEEKEND_ENTRY_BG_HOVER = "#223e60"
+
+# Time dropdown values (5-min steps, 00:00 - 23:55)
+TIME_VALUES = [f"{h:02d}:{m:02d}" for h in range(24) for m in range(0, 60, 5)]
+PAUSE_VALUES = [str(m) for m in range(0, 125, 5)]
+
+
+def apply_combobox_style(dialog):
+    style = ttk.Style(dialog)
+    style.theme_use("clam")
+    style.configure(
+        "Dark.TCombobox",
+        fieldbackground=CELL_BG, background=CELL_BG,
+        foreground=TEXT, arrowcolor=ACCENT,
+        bordercolor=TEXT_MUTED, lightcolor=CELL_BG, darkcolor=CELL_BG,
+        selectbackground=ENTRY_BG, selectforeground=TEXT,
+    )
+    style.map("Dark.TCombobox",
+        fieldbackground=[("readonly", CELL_BG)],
+        selectbackground=[("readonly", CELL_BG)],
+        selectforeground=[("readonly", TEXT)],
+        bordercolor=[("focus", ACCENT)],
+    )
+    dialog.option_add("*TCombobox*Listbox.background", CELL_BG)
+    dialog.option_add("*TCombobox*Listbox.foreground", TEXT)
+    dialog.option_add("*TCombobox*Listbox.selectBackground", ACCENT)
+    dialog.option_add("*TCombobox*Listbox.selectForeground", "#ffffff")
+    dialog.option_add("*TCombobox*Listbox.font", FONT)
