+++ src/ui.py
@@ -1,78 +1,32 @@
 # src/ui.py
 import tkinter as tk
-from tkinter import messagebox, ttk
+from tkinter import messagebox
 import calendar
 import ctypes
 import datetime
 import os
 import platform
-import sys
 import threading
 import traceback
-from src.storage import Storage
-from src.time_utils import calculate_hours, validate_entry, get_week_dates, get_week_label, week_spans_months
-from src.report import generate_report, generate_pdf
-from src.mail import (
-    get_gmail_service,
-    send_email,
-    refresh_token_if_needed,
-    TokenAuthError,
-    TokenNetworkError,
-)
-from src.autostart import enable_autostart, disable_autostart
-from src.platform_open import open_folder
+from src.time_utils import calculate_hours, get_week_dates, get_week_label, week_spans_months
+from src.mail import refresh_token_if_needed, TokenAuthError, TokenNetworkError
 from src.version import VERSION
+
+from src.dialogs.entry_dialog import open_entry_dialog
+from src.dialogs.send_dialog import open_send_dialog
+from src.dialogs.settings_dialog import open_settings_dialog
+from src.theme import (
+    BG, CELL_BG, WEEKEND_BG, ACCENT, TEXT, TEXT_MUTED,
+    ENTRY_BG, WEEKEND_ENTRY_BG, WEEKEND_FG,
+    FONT, FONT_SMALL, FONT_BOLD, FONT_HEADER, FONT_FOOTER,
+    CELL_BG_HOVER, WEEKEND_BG_HOVER, ENTRY_BG_HOVER, WEEKEND_ENTRY_BG_HOVER,
+)
 
 DAYS_DE = ["Mo", "Di", "Mi", "Do", "Fr", "Sa", "So"]
 MONTHS_DE = [
     "", "Januar", "Februar", "März", "April", "Mai", "Juni",
     "Juli", "August", "September", "Oktober", "November", "Dezember"
 ]
-
-# Dark Modern color palette
-BG = "#1a1a2e"
-CELL_BG = "#16213e"
-WEEKEND_BG = "#0f3460"
-ACCENT = "#e94560"
-STATUS_OK = "#4ade80"
-TEXT = "#e0e0e0"
-TEXT_MUTED = "#888888"
-ENTRY_BG = "#1a3a5c"
-WEEKEND_ENTRY_BG = "#1a3050"
-FONT = ("Segoe UI", 10)
-FONT_SMALL = ("Segoe UI", 8)
-FONT_BOLD = ("Segoe UI", 10, "bold")
-FONT_HEADER = ("Segoe UI", 16, "bold")
-FONT_FOOTER = ("Segoe UI", 12, "bold")
-
-# Hover colors (slightly lighter variants)
-CELL_BG_HOVER = "#1e2d52"
-WEEKEND_BG_HOVER = "#153a6e"
-ENTRY_BG_HOVER = "#224a70"
-WEEKEND_ENTRY_BG_HOVER = "#223e60"
-
-# Time dropdown values (5-min steps, 00:00 - 23:55)
-TIME_VALUES = [f"{h:02d}:{m:02d}" for h in range(24) for m in range(0, 60, 5)]
-
-
-PAUSE_VALUES = [str(m) for m in range(0, 125, 5)]
-
-
-def resolve_autostart_target(base_path):
-    """Return (target, arguments) for the current runtime/platform.
-
-    Frozen Windows/macOS: the executable itself.
-    Frozen Linux: $APPIMAGE if set (persistent path), otherwise sys.executable.
-    Script mode: Python interpreter + main.py.
-    """
-    if getattr(sys, "frozen", False):
-        if platform.system() == "Linux":
-            target = os.environ.get("APPIMAGE") or sys.executable
-        else:
-            target = sys.executable
-        return target, "--minimized"
-    main_py = os.path.join(base_path, "src", "main.py")
-    return sys.executable, f"{main_py} --minimized"
 
 
 class App:
@@ -150,7 +104,7 @@ class App:
         frame.pack(fill=tk.X, padx=10, pady=(10, 0))
 
         tk.Button(
-            frame, text="\u2039", command=self._prev, width=3,
+            frame, text="‹", command=self._prev, width=3,
             font=FONT_BOLD, bg=CELL_BG, fg=ACCENT,
             activebackground=ENTRY_BG, activeforeground=ACCENT,
             relief=tk.FLAT, cursor="hand2"
@@ -182,14 +136,14 @@ class App:
         self.header_label.pack(side=tk.LEFT, expand=True)
 
         tk.Button(
-            frame, text="\u2699", command=self._open_settings, width=3,
+            frame, text="⚙", command=self._open_settings, width=3,
             font=FONT_BOLD, bg=CELL_BG, fg=TEXT_MUTED,
             activebackground=ENTRY_BG, activeforeground=TEXT,
             relief=tk.FLAT, cursor="hand2"
         ).pack(side=tk.RIGHT)
 
         tk.Button(
-            frame, text="\u203a", command=self._next, width=3,
+            frame, text="›", command=self._next, width=3,
             font=FONT_BOLD, bg=CELL_BG, fg=ACCENT,
             activebackground=ENTRY_BG, activeforeground=ACCENT,
             relief=tk.FLAT, cursor="hand2"
@@ -210,7 +164,7 @@ class App:
         self.footer_label.pack(side=tk.LEFT, expand=True)
 
         tk.Button(
-            footer_frame, text="Monat senden", command=self._send_report,
+            footer_frame, text="Monat senden", command=self._send,
             font=FONT, bg=CELL_BG, fg=TEXT,
             activebackground=ENTRY_BG, activeforeground=TEXT,
             relief=tk.FLAT, padx=12, pady=4, cursor="hand2"
@@ -269,304 +223,11 @@ class App:
             self.btn_week.config(bg=ACCENT, fg="#ffffff", activebackground=ACCENT, activeforeground="#ffffff")
             self.btn_month.config(bg=CELL_BG, fg=TEXT_MUTED, activebackground=ENTRY_BG, activeforeground=TEXT)
 
-    def _apply_combobox_style(self, dialog):
-        style = ttk.Style(dialog)
-        style.theme_use("clam")
-        style.configure(
-            "Dark.TCombobox",
-            fieldbackground=CELL_BG, background=CELL_BG,
-            foreground=TEXT, arrowcolor=ACCENT,
-            bordercolor=TEXT_MUTED, lightcolor=CELL_BG, darkcolor=CELL_BG,
-            selectbackground=ENTRY_BG, selectforeground=TEXT
-        )
-        style.map("Dark.TCombobox",
-            fieldbackground=[("readonly", CELL_BG)],
-            selectbackground=[("readonly", CELL_BG)],
-            selectforeground=[("readonly", TEXT)],
-            bordercolor=[("focus", ACCENT)],
-        )
-        dialog.option_add("*TCombobox*Listbox.background", CELL_BG)
-        dialog.option_add("*TCombobox*Listbox.foreground", TEXT)
-        dialog.option_add("*TCombobox*Listbox.selectBackground", ACCENT)
-        dialog.option_add("*TCombobox*Listbox.selectForeground", "#ffffff")
-        dialog.option_add("*TCombobox*Listbox.font", FONT)
-
     def _open_settings(self):
-        dialog = tk.Toplevel(self.root)
-        dialog.title("Einstellungen")
-        dialog.resizable(False, False)
-        dialog.grab_set()
-        dialog.configure(bg=BG)
-
-        self._apply_combobox_style(dialog)
-
-        # Gmail-Zugangsdaten section
-        tk.Label(
-            dialog, text="— Gmail-Zugangsdaten —", font=FONT_BOLD,
-            bg=BG, fg=TEXT_MUTED
-        ).grid(row=0, column=0, columnspan=2, padx=10, pady=(8, 4))
-
-        creds_path = os.path.join(self.base_path, "credentials.json")
-
-        tk.Label(
-            dialog, text="Datenordner:", font=FONT, bg=BG, fg=TEXT
-        ).grid(row=1, column=0, padx=10, pady=4, sticky="w")
-
-        creds_row = tk.Frame(dialog, bg=BG)
-        creds_row.grid(row=1, column=1, padx=10, pady=4, sticky="w")
-
-        def open_data_folder():
-            try:
-                open_folder(self.base_path)
-            except Exception as e:
-                messagebox.showerror(
-                    "Ordner konnte nicht geöffnet werden",
-                    f"{type(e).__name__}: {e}\n\n{traceback.format_exc()}",
-                    parent=dialog,
-                )
-
-        tk.Button(
-            creds_row, text="Ordner öffnen", command=open_data_folder,
-            font=FONT, bg=CELL_BG, fg=TEXT,
-            activebackground=ENTRY_BG, activeforeground=TEXT,
-            relief=tk.FLAT, padx=12, pady=2, cursor="hand2"
-        ).pack(side=tk.LEFT)
-
-        status_label = tk.Label(
-            creds_row, text="", font=FONT_SMALL, bg=BG
+        open_settings_dialog(
+            self.root, self.settings, self.base_path,
+            on_change=self._refresh,
         )
-        status_label.pack(side=tk.LEFT, padx=(10, 0))
-
-        def refresh_status():
-            if not status_label.winfo_exists():
-                return
-            if os.path.exists(creds_path):
-                status_label.config(
-                    text="✓ credentials.json vorhanden", fg=STATUS_OK
-                )
-            else:
-                status_label.config(
-                    text="✗ credentials.json fehlt", fg=ACCENT
-                )
-            dialog.after(500, refresh_status)
-
-        refresh_status()
-
-        # Email
-        tk.Label(
-            dialog, text="Absender:", font=FONT, bg=BG, fg=TEXT
-        ).grid(row=2, column=0, padx=10, pady=8, sticky="w")
-
-        email_var = tk.StringVar(value=self.settings.get("email"))
-        tk.Entry(
-            dialog, textvariable=email_var, width=25, font=FONT,
-            bg=CELL_BG, fg=TEXT, insertbackground=ACCENT,
-            relief=tk.FLAT, highlightbackground=TEXT_MUTED,
-            highlightcolor=ACCENT, highlightthickness=1
-        ).grid(row=2, column=1, padx=10, pady=8)
-
-        # Default start time
-        tk.Label(
-            dialog, text="Standard-Start:", font=FONT, bg=BG, fg=TEXT
-        ).grid(row=3, column=0, padx=10, pady=8, sticky="w")
-
-        start_var = tk.StringVar(value=self.settings.get("default_start"))
-        ttk.Combobox(
-            dialog, textvariable=start_var, values=TIME_VALUES,
-            width=8, font=FONT, style="Dark.TCombobox", state="readonly"
-        ).grid(row=3, column=1, padx=10, pady=8)
-
-        # Default end time
-        tk.Label(
-            dialog, text="Standard-Ende:", font=FONT, bg=BG, fg=TEXT
-        ).grid(row=4, column=0, padx=10, pady=8, sticky="w")
-
-        end_var = tk.StringVar(value=self.settings.get("default_end"))
-        ttk.Combobox(
-            dialog, textvariable=end_var, values=TIME_VALUES,
-            width=8, font=FONT, style="Dark.TCombobox", state="readonly"
-        ).grid(row=4, column=1, padx=10, pady=8)
-
-        # Default pause
-        tk.Label(
-            dialog, text="Standard-Pause (Min):", font=FONT, bg=BG, fg=TEXT
-        ).grid(row=5, column=0, padx=10, pady=8, sticky="w")
-
-        pause_var = tk.StringVar(value=str(self.settings.get("default_pause")))
-        ttk.Combobox(
-            dialog, textvariable=pause_var, values=PAUSE_VALUES,
-            width=8, font=FONT, style="Dark.TCombobox", state="readonly"
-        ).grid(row=5, column=1, padx=10, pady=8)
-
-        # Recipient
-        tk.Label(
-            dialog, text="Empfänger:", font=FONT, bg=BG, fg=TEXT
-        ).grid(row=6, column=0, padx=10, pady=8, sticky="w")
-
-        recipient_var = tk.StringVar(value=self.settings.get("recipient"))
-        tk.Entry(
-            dialog, textvariable=recipient_var, width=25, font=FONT,
-            bg=CELL_BG, fg=TEXT, insertbackground=ACCENT,
-            relief=tk.FLAT, highlightbackground=TEXT_MUTED,
-            highlightcolor=ACCENT, highlightthickness=1
-        ).grid(row=6, column=1, padx=10, pady=8)
-
-        # Name
-        tk.Label(
-            dialog, text="Name:", font=FONT, bg=BG, fg=TEXT
-        ).grid(row=7, column=0, padx=10, pady=8, sticky="w")
-
-        name_var = tk.StringVar(value=self.settings.get("name"))
-        tk.Entry(
-            dialog, textvariable=name_var, width=25, font=FONT,
-            bg=CELL_BG, fg=TEXT, insertbackground=ACCENT,
-            relief=tk.FLAT, highlightbackground=TEXT_MUTED,
-            highlightcolor=ACCENT, highlightthickness=1
-        ).grid(row=7, column=1, padx=10, pady=8)
-
-        # Stundenlohn (optional)
-        tk.Label(
-            dialog, text="Stundenlohn (€):", font=FONT, bg=BG, fg=TEXT
-        ).grid(row=8, column=0, padx=10, pady=8, sticky="w")
-
-        rate_var = tk.StringVar(value=str(self.settings.get("hourly_rate") or ""))
-        tk.Entry(
-            dialog, textvariable=rate_var, width=10, font=FONT,
-            bg=CELL_BG, fg=TEXT, insertbackground=ACCENT,
-            relief=tk.FLAT, highlightbackground=TEXT_MUTED,
-            highlightcolor=ACCENT, highlightthickness=1
-        ).grid(row=8, column=1, padx=10, pady=8, sticky="w")
-
-        tk.Label(
-            dialog, text="(optional – nur für dich sichtbar)", font=FONT_SMALL,
-            bg=BG, fg=TEXT_MUTED
-        ).grid(row=8, column=1, padx=(120, 10), pady=8, sticky="w")
-
-        # Mail-Vorlage
-        tk.Label(
-            dialog, text="— Mail-Vorlage —", font=FONT_BOLD, bg=BG, fg=TEXT_MUTED
-        ).grid(row=9, column=0, columnspan=2, padx=10, pady=(16, 4))
-
-        tk.Label(
-            dialog, text="Betreff:", font=FONT, bg=BG, fg=TEXT
-        ).grid(row=10, column=0, padx=10, pady=4, sticky="w")
-        subject_var = tk.StringVar(value=self.settings.get("mail_subject"))
-        tk.Entry(
-            dialog, textvariable=subject_var, width=35, font=FONT,
-            bg=CELL_BG, fg=TEXT, insertbackground=ACCENT,
-            relief=tk.FLAT, highlightbackground=TEXT_MUTED,
-            highlightcolor=ACCENT, highlightthickness=1
-        ).grid(row=10, column=1, padx=10, pady=4)
-
-        tk.Label(
-            dialog, text="Anrede:", font=FONT, bg=BG, fg=TEXT
-        ).grid(row=11, column=0, padx=10, pady=4, sticky="w")
-        greeting_var = tk.StringVar(value=self.settings.get("mail_greeting"))
-        tk.Entry(
-            dialog, textvariable=greeting_var, width=35, font=FONT,
-            bg=CELL_BG, fg=TEXT, insertbackground=ACCENT,
-            relief=tk.FLAT, highlightbackground=TEXT_MUTED,
-            highlightcolor=ACCENT, highlightthickness=1
-        ).grid(row=11, column=1, padx=10, pady=4)
-
-        tk.Label(
-            dialog, text="Inhalt:", font=FONT, bg=BG, fg=TEXT
-        ).grid(row=12, column=0, padx=10, pady=4, sticky="nw")
-        content_text = tk.Text(
-            dialog, width=35, height=3, font=FONT,
-            bg=CELL_BG, fg=TEXT, insertbackground=ACCENT,
-            relief=tk.FLAT, highlightbackground=TEXT_MUTED,
-            highlightcolor=ACCENT, highlightthickness=1, wrap=tk.WORD
-        )
-        content_text.grid(row=12, column=1, padx=10, pady=4)
-        content_text.insert("1.0", self.settings.get("mail_content"))
-
-        tk.Label(
-            dialog, text="Gruß:", font=FONT, bg=BG, fg=TEXT
-        ).grid(row=13, column=0, padx=10, pady=4, sticky="nw")
-        closing_text = tk.Text(
-            dialog, width=35, height=2, font=FONT,
-            bg=CELL_BG, fg=TEXT, insertbackground=ACCENT,
-            relief=tk.FLAT, highlightbackground=TEXT_MUTED,
-            highlightcolor=ACCENT, highlightthickness=1, wrap=tk.WORD
-        )
-        closing_text.grid(row=13, column=1, padx=10, pady=4)
-        closing_text.insert("1.0", self.settings.get("mail_closing"))
-
-        tk.Label(
-            dialog, text="Platzhalter: {zeitraum}, {gesamt}", font=("Segoe UI", 8),
-            bg=BG, fg=TEXT_MUTED
-        ).grid(row=14, column=0, columnspan=2, padx=10, pady=(0, 4))
-
-        # Autostart
-        autostart_var = tk.BooleanVar(value=self.settings.get("autostart"))
-        tk.Checkbutton(
-            dialog, text="Autostart (minimiert bei Anmeldung)",
-            variable=autostart_var, font=FONT,
-            bg=BG, fg=TEXT, selectcolor=CELL_BG,
-            activebackground=BG, activeforeground=TEXT,
-            cursor="hand2"
-        ).grid(row=15, column=0, columnspan=2, padx=10, pady=8, sticky="w")
-
-        def save_settings():
-            ok, msg = validate_entry(start_var.get(), end_var.get())
-            if not ok:
-                messagebox.showerror("Standard-Arbeitszeit ungültig", msg, parent=dialog)
-                return
-
-            new_autostart = autostart_var.get()
-            old_autostart = self.settings.get("autostart")
-
-            if new_autostart != old_autostart:
-                try:
-                    if new_autostart:
-                        target, arguments = resolve_autostart_target(self.base_path)
-                        enable_autostart(target, arguments)
-                    else:
-                        disable_autostart()
-                    self.settings.set("autostart", new_autostart)
-                except Exception as e:
-                    messagebox.showerror(
-                        "Autostart-Fehler",
-                        f"Autostart konnte nicht geändert werden:\n{e}",
-                        parent=dialog,
-                    )
-                    return
-
-            self.settings.set("email", email_var.get())
-            self.settings.set("default_start", start_var.get())
-            self.settings.set("default_end", end_var.get())
-            self.settings.set("default_pause", int(pause_var.get()))
-            self.settings.set("recipient", recipient_var.get())
-            self.settings.set("name", name_var.get())
-            self.settings.set("mail_subject", subject_var.get())
-            self.settings.set("mail_greeting", greeting_var.get())
-            self.settings.set("mail_content", content_text.get("1.0", "end-1c"))
-            self.settings.set("mail_closing", closing_text.get("1.0", "end-1c"))
-            rate_str = rate_var.get().strip()
-            try:
-                self.settings.set("hourly_rate", float(rate_str) if rate_str else 0.0)
-            except ValueError:
-                self.settings.set("hourly_rate", 0.0)
-            self._refresh()
-            dialog.destroy()
-
-        btn_frame = tk.Frame(dialog, bg=BG)
-        btn_frame.grid(row=16, column=0, columnspan=2, pady=12)
-
-        tk.Button(
-            btn_frame, text="Speichern", command=save_settings, font=FONT_BOLD,
-            bg=ACCENT, fg="#ffffff",
-            activebackground="#c73550", activeforeground="#ffffff",
-            relief=tk.FLAT, padx=16, pady=4, cursor="hand2"
-        ).pack(side=tk.LEFT, padx=5)
-
-        tk.Button(
-            btn_frame, text="Abbrechen", command=dialog.destroy, font=FONT,
-            bg=CELL_BG, fg=TEXT,
-            activebackground=ENTRY_BG, activeforeground=TEXT,
-            relief=tk.FLAT, padx=16, pady=4, cursor="hand2"
-        ).pack(side=tk.LEFT, padx=5)
 
     def _refresh(self):
         if self.view_mode == "month":
@@ -587,7 +248,7 @@ class App:
 
         # Column headers
         for col, day_name in enumerate(DAYS_DE):
-            fg = TEXT_MUTED if col < 5 else "#6c6c80"
+            fg = TEXT_MUTED if col < 5 else WEEKEND_FG
             lbl = tk.Label(
                 new_frame, text=day_name, font=FONT_BOLD,
                 bg=BG, fg=fg
@@ -615,7 +276,7 @@ class App:
                 text = str(day)
                 fg = TEXT
                 if is_weekend and not entry:
-                    fg = "#6c6c80"
+                    fg = WEEKEND_FG
 
                 if entry:
                     pause = entry.get("pause", 0)
@@ -680,7 +341,7 @@ class App:
 
         # Column headers
         for col, day_name in enumerate(DAYS_DE):
-            fg = TEXT_MUTED if col < 5 else "#6c6c80"
+            fg = TEXT_MUTED if col < 5 else WEEKEND_FG
             lbl = tk.Label(
                 new_frame, text=day_name, font=FONT_BOLD,
                 bg=BG, fg=fg
@@ -700,7 +361,7 @@ class App:
 
             fg = TEXT
             if is_weekend and not entry:
-                fg = "#6c6c80"
+                fg = WEEKEND_FG
 
             if entry:
                 pause = entry.get("pause", 0)
@@ -773,332 +434,10 @@ class App:
             self._refresh()
 
     def _open_dialog(self, date_str):
-        dialog = tk.Toplevel(self.root)
-        dialog.title(date_str)
-        dialog.resizable(False, False)
-        dialog.grab_set()
-        dialog.configure(bg=BG)
-
-        entry = self.storage.get(date_str)
-
-        self._apply_combobox_style(dialog)
-
-        tk.Label(
-            dialog, text="Start:", font=FONT,
-            bg=BG, fg=TEXT
-        ).grid(row=0, column=0, padx=10, pady=8, sticky="w")
-
-        start_var = tk.StringVar(value=entry["start"] if entry else self.settings.get("default_start"))
-        start_cb = ttk.Combobox(
-            dialog, textvariable=start_var, values=TIME_VALUES,
-            width=8, font=FONT, style="Dark.TCombobox", state="readonly"
+        open_entry_dialog(
+            self.root, date_str, self.storage, self.settings,
+            on_change=self._refresh,
         )
-        start_cb.grid(row=0, column=1, padx=10, pady=8)
 
-        tk.Label(
-            dialog, text="Ende:", font=FONT,
-            bg=BG, fg=TEXT
-        ).grid(row=1, column=0, padx=10, pady=8, sticky="w")
-
-        end_var = tk.StringVar(value=entry["end"] if entry else self.settings.get("default_end"))
-        end_cb = ttk.Combobox(
-            dialog, textvariable=end_var, values=TIME_VALUES,
-            width=8, font=FONT, style="Dark.TCombobox", state="readonly"
-        )
-        end_cb.grid(row=1, column=1, padx=10, pady=8)
-
-        # Pause dropdown
-        tk.Label(
-            dialog, text="Pause (Min):", font=FONT,
-            bg=BG, fg=TEXT
-        ).grid(row=2, column=0, padx=10, pady=8, sticky="w")
-
-        default_pause = self.settings.get("default_pause")
-        if entry and "pause" in entry:
-            current_pause = str(entry["pause"])
-        else:
-            current_pause = str(default_pause) if not entry else "0"
-        pause_var = tk.StringVar(value=current_pause)
-
-        pause_cb = ttk.Combobox(
-            dialog, textvariable=pause_var, values=PAUSE_VALUES,
-            width=8, font=FONT, style="Dark.TCombobox", state="readonly"
-        )
-        pause_cb.grid(row=2, column=1, padx=10, pady=8)
-
-        def save():
-            ok, msg = validate_entry(start_var.get(), end_var.get(), pause_minutes=int(pause_var.get()))
-            if not ok:
-                messagebox.showerror("Fehler", msg, parent=dialog)
-                return
-            self.storage.save(date_str, start_var.get(), end_var.get(), pause=int(pause_var.get()))
-            dialog.destroy()
-            self._refresh()
-
-        def delete():
-            self.storage.delete(date_str)
-            dialog.destroy()
-            self._refresh()
-
-        btn_frame = tk.Frame(dialog, bg=BG)
-        btn_frame.grid(row=3, column=0, columnspan=2, pady=12)
-
-        tk.Button(
-            btn_frame, text="Speichern", command=save, font=FONT_BOLD,
-            bg=ACCENT, fg="#ffffff",
-            activebackground="#c73550", activeforeground="#ffffff",
-            relief=tk.FLAT, padx=16, pady=4, cursor="hand2"
-        ).pack(side=tk.LEFT, padx=5)
-
-        tk.Button(
-            btn_frame, text="Löschen", command=delete, font=FONT,
-            bg=CELL_BG, fg=TEXT,
-            activebackground=ENTRY_BG, activeforeground=TEXT,
-            relief=tk.FLAT, padx=16, pady=4, cursor="hand2"
-        ).pack(side=tk.LEFT, padx=5)
-
-    def _show_missing_credentials_dialog(self):
-        dialog = tk.Toplevel(self.root)
-        dialog.title("Keine Zugangsdaten")
-        dialog.resizable(False, False)
-        dialog.grab_set()
-        dialog.configure(bg=BG)
-
-        tk.Label(
-            dialog,
-            text=(
-                "credentials.json nicht gefunden.\n\n"
-                "Bitte erstelle ein Google Cloud Projekt mit Gmail API "
-                "und lade die OAuth2 Client-ID als credentials.json in "
-                "den Datenordner."
-            ),
-            font=FONT, bg=BG, fg=TEXT,
-            wraplength=380, justify="left",
-        ).grid(row=0, column=0, columnspan=2, padx=20, pady=(20, 12))
-
-        def open_and_close():
-            try:
-                open_folder(self.base_path)
-            except Exception as e:
-                messagebox.showerror(
-                    "Ordner konnte nicht geöffnet werden",
-                    f"{type(e).__name__}: {e}\n\n{traceback.format_exc()}",
-                    parent=dialog,
-                )
-                return
-            dialog.destroy()
-
-        btn_frame = tk.Frame(dialog, bg=BG)
-        btn_frame.grid(row=1, column=0, columnspan=2, pady=(0, 16))
-
-        tk.Button(
-            btn_frame, text="Datenordner öffnen", command=open_and_close,
-            font=FONT_BOLD, bg=ACCENT, fg="#ffffff",
-            activebackground="#c73550", activeforeground="#ffffff",
-            relief=tk.FLAT, padx=16, pady=4, cursor="hand2",
-        ).pack(side=tk.LEFT, padx=5)
-
-        tk.Button(
-            btn_frame, text="OK", command=dialog.destroy,
-            font=FONT, bg=CELL_BG, fg=TEXT,
-            activebackground=ENTRY_BG, activeforeground=TEXT,
-            relief=tk.FLAT, padx=16, pady=4, cursor="hand2",
-        ).pack(side=tk.LEFT, padx=5)
-
-    def _send_report(self):
-        recipient = self.settings.get("recipient")
-        if not recipient:
-            messagebox.showwarning(
-                "Kein Empfänger",
-                "Bitte zuerst einen Empfänger in den Einstellungen angeben.",
-                parent=self.root
-            )
-            return
-
-        credentials_path = os.path.join(self.base_path, "credentials.json")
-        token_path = os.path.join(self.base_path, "token.json")
-
-        if not os.path.exists(credentials_path):
-            self._show_missing_credentials_dialog()
-            return
-
-        # Date range dialog
-        dialog = tk.Toplevel(self.root)
-        dialog.title("Zeitraum wählen")
-        dialog.resizable(False, False)
-        dialog.grab_set()
-        dialog.configure(bg=BG)
-
-        self._apply_combobox_style(dialog)
-
-        today = datetime.date.today()
-        # Default: one month ago to today
-        if today.month == 1:
-            from_default = today.replace(year=today.year - 1, month=12)
-        else:
-            from_month = today.month - 1
-            from_year = today.year
-            max_day = calendar.monthrange(from_year, from_month)[1]
-            from_default = today.replace(month=from_month, day=min(today.day, max_day))
-        month_values = [str(m) for m in range(1, 13)]
-        year_values = [str(y) for y in range(2020, today.year + 2)]
-
-        def update_day_values(day_cb, day_var, month_var, year_var):
-            """Update day combobox values based on selected month/year."""
-            try:
-                m = int(month_var.get())
-                y = int(year_var.get())
-                max_day = calendar.monthrange(y, m)[1]
-            except (ValueError, KeyError):
-                max_day = 31
-            new_values = [str(d) for d in range(1, max_day + 1)]
-            day_cb["values"] = new_values
-            if int(day_var.get()) > max_day:
-                day_var.set(str(max_day))
-
-        # Von
-        tk.Label(
-            dialog, text="Von:", font=FONT, bg=BG, fg=TEXT
-        ).grid(row=0, column=0, padx=(10, 5), pady=8, sticky="w")
-
-        from_day_var = tk.StringVar(value=str(from_default.day))
-        from_max = calendar.monthrange(from_default.year, from_default.month)[1]
-        from_day_cb = ttk.Combobox(
-            dialog, textvariable=from_day_var,
-            values=[str(d) for d in range(1, from_max + 1)],
-            width=3, font=FONT, style="Dark.TCombobox", state="readonly"
-        )
-        from_day_cb.grid(row=0, column=1, padx=2, pady=8)
-
-        tk.Label(dialog, text=".", font=FONT, bg=BG, fg=TEXT).grid(row=0, column=2)
-
-        from_month_var = tk.StringVar(value=str(from_default.month))
-        from_month_cb = ttk.Combobox(
-            dialog, textvariable=from_month_var, values=month_values,
-            width=3, font=FONT, style="Dark.TCombobox", state="readonly"
-        )
-        from_month_cb.grid(row=0, column=3, padx=2, pady=8)
-
-        tk.Label(dialog, text=".", font=FONT, bg=BG, fg=TEXT).grid(row=0, column=4)
-
-        from_year_var = tk.StringVar(value=str(from_default.year))
-        from_year_cb = ttk.Combobox(
-            dialog, textvariable=from_year_var, values=year_values,
-            width=5, font=FONT, style="Dark.TCombobox", state="readonly"
-        )
-        from_year_cb.grid(row=0, column=5, padx=(2, 10), pady=8)
-
-        from_month_var.trace_add("write", lambda *_: update_day_values(from_day_cb, from_day_var, from_month_var, from_year_var))
-        from_year_var.trace_add("write", lambda *_: update_day_values(from_day_cb, from_day_var, from_month_var, from_year_var))
-
-        # Bis
-        tk.Label(
-            dialog, text="Bis:", font=FONT, bg=BG, fg=TEXT
-        ).grid(row=1, column=0, padx=(10, 5), pady=8, sticky="w")
-
-        to_day_var = tk.StringVar(value=str(today.day))
-        to_max = calendar.monthrange(today.year, today.month)[1]
-        to_day_cb = ttk.Combobox(
-            dialog, textvariable=to_day_var,
-            values=[str(d) for d in range(1, to_max + 1)],
-            width=3, font=FONT, style="Dark.TCombobox", state="readonly"
-        )
-        to_day_cb.grid(row=1, column=1, padx=2, pady=8)
-
-        tk.Label(dialog, text=".", font=FONT, bg=BG, fg=TEXT).grid(row=1, column=2)
-
-        to_month_var = tk.StringVar(value=str(today.month))
-        to_month_cb = ttk.Combobox(
-            dialog, textvariable=to_month_var, values=month_values,
-            width=3, font=FONT, style="Dark.TCombobox", state="readonly"
-        )
-        to_month_cb.grid(row=1, column=3, padx=2, pady=8)
-
-        tk.Label(dialog, text=".", font=FONT, bg=BG, fg=TEXT).grid(row=1, column=4)
-
-        to_year_var = tk.StringVar(value=str(today.year))
-        to_year_cb = ttk.Combobox(
-            dialog, textvariable=to_year_var, values=year_values,
-            width=5, font=FONT, style="Dark.TCombobox", state="readonly"
-        )
-        to_year_cb.grid(row=1, column=5, padx=(2, 10), pady=8)
-
-        to_month_var.trace_add("write", lambda *_: update_day_values(to_day_cb, to_day_var, to_month_var, to_year_var))
-        to_year_var.trace_add("write", lambda *_: update_day_values(to_day_cb, to_day_var, to_month_var, to_year_var))
-
-        def do_send():
-            try:
-                date_from = datetime.date(
-                    int(from_year_var.get()), int(from_month_var.get()), int(from_day_var.get())
-                )
-                date_to = datetime.date(
-                    int(to_year_var.get()), int(to_month_var.get()), int(to_day_var.get())
-                )
-            except ValueError:
-                messagebox.showerror("Ungültiges Datum", "Bitte ein gültiges Datum eingeben.", parent=dialog)
-                return
-
-            if date_from > date_to:
-                messagebox.showerror("Ungültiger Zeitraum", "Das Von-Datum muss vor dem Bis-Datum liegen.", parent=dialog)
-                return
-
-            entries = self.storage.get_all()
-
-            greeting = self.settings.get("mail_greeting")
-            content = self.settings.get("mail_content")
-            closing = self.settings.get("mail_closing")
-
-            html, total = generate_report(
-                date_from, date_to, entries,
-                greeting=greeting, content=content, closing=closing
-            )
-
-            if html is None:
-                messagebox.showinfo(
-                    "Keine Einträge",
-                    f"Keine Einträge für {date_from.strftime('%d.%m.%Y')} – {date_to.strftime('%d.%m.%Y')} vorhanden.",
-                    parent=dialog
-                )
-                return
-
-            label = f"{date_from.strftime('%d.%m.%Y')} – {date_to.strftime('%d.%m.%Y')}"
-
-            try:
-                pdf_bytes = generate_pdf(date_from, date_to, entries, name=self.settings.get("name"))
-                service = get_gmail_service(credentials_path, token_path)
-                subject_template = self.settings.get("mail_subject")
-                subject = subject_template.replace("{zeitraum}", label).replace("{gesamt}", f"{total}h")
-                pdf_filename = f"Zeiterfassung_{date_from.strftime('%Y%m%d')}_{date_to.strftime('%Y%m%d')}.pdf"
-                send_email(service, recipient, subject, html,
-                           pdf_bytes=pdf_bytes, pdf_filename=pdf_filename)
-                dialog.destroy()
-                messagebox.showinfo(
-                    "Gesendet",
-                    f"Bericht für {date_from.strftime('%d.%m.%Y')} – {date_to.strftime('%d.%m.%Y')} wurde an {recipient} gesendet.",
-                    parent=self.root
-                )
-            except FileNotFoundError as e:
-                messagebox.showerror("Fehler", str(e), parent=dialog)
-            except Exception as e:
-                messagebox.showerror(
-                    "Senden fehlgeschlagen",
-                    f"{type(e).__name__}: {e}\n\n{traceback.format_exc()}",
-                    parent=dialog
-                )
-
-        btn_frame = tk.Frame(dialog, bg=BG)
-        btn_frame.grid(row=2, column=0, columnspan=6, pady=12)
-
-        tk.Button(
-            btn_frame, text="Senden", command=do_send, font=FONT_BOLD,
-            bg=ACCENT, fg="#ffffff",
-            activebackground="#c73550", activeforeground="#ffffff",
-            relief=tk.FLAT, padx=16, pady=4, cursor="hand2"
-        ).pack(side=tk.LEFT, padx=5)
-
-        tk.Button(
-            btn_frame, text="Abbrechen", command=dialog.destroy, font=FONT,
-            bg=CELL_BG, fg=TEXT,
-            activebackground=ENTRY_BG, activeforeground=TEXT,
-            relief=tk.FLAT, padx=16, pady=4, cursor="hand2"
-        ).pack(side=tk.LEFT, padx=5)
+    def _send(self):
+        open_send_dialog(self.root, self.storage, self.settings, self.base_path)
